Deposit + withdrawl constructed (withdrawl has a bit of an issue)

+++ LogicData/Bank.cs
@@ -222,7 +222,7 @@ namespace WorldBankApp.LogicData
         }
 
         // Withdraw
-        public void Withdraw(int withdraw)
+        public void Withdraw(double withdraw)
         {
             if (activeAccount != null)
             {
